pyproject.toml
[tool.poetry]
name = "nonebot-plugin-reminder"
version = "0.3.0-1"
description = "This plugin can remind you not to forget something"
authors = ["velor2012 <38395332+velor2012@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot-adapter-onebot = "^2.3.1"
chinese-calendar = "^1.9.0"
nonebot2 = "^2.1.3"
nonebot-plugin-apscheduler = "^0.3.0"
nonebot-plugin-localstore = "^0.5.1"
nonebot-plugin-send-anything-anywhere = "^0.5.0"
aiofiles = "^23.2.1"
ujson = "^5.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
